# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SDSS API",
        "description": "Secure Document Sharing System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Local Dev"
        }
    ],
    "paths": {
...
        "/api/folders": {
            "get": {
                "tags": [
                    "Folders"
                ],
                "summary": "List root folders",
                "operationId": "0f3128f67b353b6cee06207d86b3b990",
                "responses": {
                    "200": {
                        "description": "Folders retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Folders"
                ],
                "summary": "Create a new folder",
                "operationId": "f2645643f162d5243eab19bceea7577f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Documents"
                                    },
                                    "parent_id": {
                                        "description": "Parent folder ID (null = root)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Folder created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/folders/{folder}": {
            "get": {
                "tags": [
                    "Folders"
                ],
                "summary": "Get folder contents (subfolders + files)",
                "operationId": "0594875a027e071feadf9db35a183584",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder contents"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Folders"
                ],
                "summary": "Delete a folder and its contents",
                "operationId": "5833856a66ed937004afa5d3b2fd9b7b",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Folders"
                ],
                "summary": "Rename a folder",
                "operationId": "82bea8428f03885e0d18df0c7ddccf44",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Renamed Folder"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Folder renamed"
                    },
                    "422": {
                        "description": "Duplicate name"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/folders/{folder}/move": {
            "patch": {
                "tags": [
                    "Folders"
                ],
                "summary": "Move folder to a new parent",
                "operationId": "dd4977d9329ee29af8294102c00741ad",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "description": "Target parent folder ID (null = root)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Folder moved"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}